+++ src/test/ExSln2/c99/gen/ex_mem_DoublePoint.c
@@ -11,6 +11,7 @@
 void ex_mem_DoublePoint_ctor(ex_mem_DoublePoint * self)
 {
     memset(self, 0, sizeof(*self));
+    self->start = finlang_mem_init(new ex_mem_XyPointU8 *());
     self->end = finlang_mem_init(new ex_mem_XyPointU8 *());
 }